Added Global Method
Added a global method to calculate the distance in kilometres from to
CGPoints just put them in as though they're lat and longs where the Y is
the Lat and the X is the long. the rest of the methods may end up being
placed into a separate class for math related items.